# packages/shared/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditLogEntry": {
        "description": "One audit row exposed in the Dashboard agent activity feed.",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "resourceId": {
            "title": "Resourceid",
            "type": "string"
          },
          "resourceType": {
            "title": "Resourcetype",
            "type": "string"
          },
          "userId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Userid"
          }
        },
        "required": [
          "id",
          "action",
          "resourceType",
          "resourceId",
          "createdAt"
        ],
        "title": "AuditLogEntry",
        "type": "object"
      },
      "AuditLogResponse": {
        "description": "``GET /audit-logs`` envelope \u2014 small bounded list, no cursor.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditLogEntry"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "AuditLogResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "McpProvidersResponse": {
        "description": "``GET /mcp/providers`` envelope (read-only in M1b).",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/suitest_api__routers__mcp_providers__McpProviderPublic"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "McpProvidersResponse",
        "type": "object"
      },
...
      "RunLogPage": {
        "description": "A cursor-paginated slice of a run's concatenated stdout/stderr text.",
        "properties": {
          "lines": {
            "items": {
              "type": "string"
            },
            "title": "Lines",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "title": "RunLogPage",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "suitest_api__routers__mcp_providers__McpProviderPublic": {
        "description": "One MCP provider row \u2014 name + transport + health only (no secrets).",
        "properties": {
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "healthStatus": {
            "default": "unknown",
            "title": "Healthstatus",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "isBundled": {
            "default": false,
            "title": "Isbundled",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "lastHealthAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lasthealthat"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "transport": {
            "title": "Transport",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "transport",
          "endpoint"
        ],
        "title": "McpProviderPublic",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "suitest_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Suitest API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/audit-logs": {
      "get": {
        "description": "List audit log entries for the workspace, optionally filtered by action prefix.",
        "operationId": "list_audit_logs_api_v1_audit_logs_get",
        "parameters": [
          {
            "description": "Action prefix filter (e.g. 'agent.*')",
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Action prefix filter (e.g. 'agent.*')",
              "title": "Action"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Audit Logs",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/api/v1/mcp/providers": {
      "get": {
        "description": "List MCP providers for the workspace (read-only in M1b; CRUD in M2).",
        "operationId": "list_mcp_providers_api_v1_mcp_providers_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpProvidersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Mcp Providers",
        "tags": [
          "mcp"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/logs": {
      "get": {
        "description": "Concatenated stdout/stderr text in step_order, paginated by line offset.",
        "operationId": "get_run_logs_api_v1_runs__run_id__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunLogPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Run Logs",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}